Read environment configuration from devicetree
In order to make systems more flexible, let's add the possibility
of reading the configuration from the devicetree if no config file
was found.

This way we can use the information about the environment passed
by U-Boot itself to determine the configuration for the environment
that is currently in use.

A file /etc/fw_env_devicemap.conf can be used to match the
information passed by U-Boot with a Linux device. If such file is
missing we try to make a guess for the correct device based on
the input from the devicetree.

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>